requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.7
pydantic-settings>=2.4

numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
pillow>=10.4
matplotlib>=3.8
tqdm>=4.66

transformers>=4.44
safetensors>=0.4.4
huggingface-hub>=0.24

torch>=2.5,<3.0
torchvision>=0.20,<0.21

pymupdf>=1.24
pywin32>=306 ; sys_platform == 'win32'

umap-learn>=0.5.6